+++ src/test/java/dev/thiagogonzaga/thrillhousebot/review/PrImprovementServiceTest.java
@@ -629,7 +629,7 @@ class PrImprovementServiceTest {
             .estimateTokens(
                 PrImproveAssistantPrompts.systemPrompt()
                     + PrSuggestionPrompts.userPrompt()
-                    + PromptTemplateEscaper.fence(" ")
+                    + PromptTemplateEscaper.fenceForBudgeting()
                     + "Title"
                     + "Body"
                     + "");